requirements.txt
numpy
matplotlib
pygame
tensorflow-gpu
keras
Pillow
opencv-python
h5py
